Componentize `Pages/Wallet/Index` - Extract views inside `Pages/Wallet/Index` into separate vue view components
