+++ resources/js/Views/Wallet/WalletPageHeader.vue
@@ -1,0 +1,16 @@
+<script setup>
+import Header from "@/Components/Header.vue";
+</script>
+
+<template>
+    <Header class="max-w-md">
+        <template #default>
+            <div class="font-semibold text-xl pb-1">Your Wallet</div>
+            <div class="text-xs font-light">You can withdraw from or deposit to your wallet account. You can also transfer funds to another player.</div>
+        </template>
+    </Header>
+</template>
+
+<style scoped>
+
+</style>
